% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampler.R
\name{accept}
\alias{accept}
\title{accept propsbm with the acceptance probability alpha}
\usage{
accept(currsbm, propsbm, Edges, Mod, logjac = 0, logu = 0, ...)
}
\arguments{
\item{currsbm}{current sbm state}

\item{propsbm}{proposed sbm state}

\item{Edges}{edges object}

\item{Mod}{model object}

\item{logjac}{log jacobian of transformation of variables}

\item{logu}{log density for auxiliary variables}

\item{...}{additional arguments to pass to loglike}
}
\value{
new currsbm
}
\description{
accept propsbm with the acceptance probability alpha
}
