% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rjmcmc.R
\name{addblock}
\alias{addblock}
\title{add a block move}
\usage{
addblock(sbm, edges, mod, rho = 1)
}
\arguments{
\item{sbm}{an sbm object}

\item{edges}{an edges object}

\item{mod}{a model object}

\item{rho}{probability of choosing to add a block}
}
\value{
an updated sbm object
}
\description{
add a block move
}
