% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{ARI}
\alias{ARI}
\title{Adjusted Rand Index}
\usage{
ARI(z, truez)
}
\arguments{
\item{z}{input vector}

\item{truez}{reference vector}
}
\value{
Adjusted Rand Index of \code{z} against \code{truez}
}
\description{
Calculate the Adjusted Rand Index between two clusterings
}
\examples{
ARI(c(1,1,2,2,3,3), c(2,2,1,1,3,3)) ## 1 - doesn't care for labels
ARI(c(1,1,2,2,3,3), c(1,1,1,1,2,2)) ## 0.444
}
