% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockmat.R
\name{blockmat.blocks}
\alias{blockmat.blocks}
\title{Block matrix}
\usage{
\method{blockmat}{blocks}(blocks, kappa)
}
\arguments{
\item{blocks}{a \code{blocks} object}

\item{kappa}{number of blocks in matrix}
}
\value{
matrix with \code{kappa} rows and a 1 at \code{(k,i)} if node \code{i} is in block \code{k} under \code{blocks}
}
\description{
converts block assignments of a \code{blocks} object to a matrix of block assignments
}
