% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockmat.R
\name{blockmat.numeric}
\alias{blockmat.numeric}
\alias{blockmat.factor}
\title{Block matrix}
\usage{
\method{blockmat}{numeric}(x, kappa)

\method{blockmat}{factor}(x, kappa)
}
\arguments{
\item{x}{a numeric-vector of node-to-block assignments}

\item{kappa}{number of blocks}
}
\value{
matrix with \code{kappa} rows and a 1 at \code{(k,i)} if node \code{i} is in block \code{k} under \code{x}
}
\description{
converts a vector of block assignments to a matrix of block assignments
}
