% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockmat.R
\name{blockmat}
\alias{blockmat}
\title{Block matrix}
\usage{
blockmat(x, ...)
}
\arguments{
\item{x}{object for dispatch}

\item{...}{additional arguments for method}
}
\value{
matrix of block assignment indicators
}
\description{
converts \code{x} to a matrix of block assignments
}
\seealso{
\code{\link{blockmat.sbm}} \code{\link{blockmat.blocks}} \code{\link{blockmat.numeric}}
}
