% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockmat.R
\name{blockmat.sbm}
\alias{blockmat.sbm}
\title{Block matrix}
\usage{
\method{blockmat}{sbm}(SBM, kappa)
}
\arguments{
\item{SBM}{an \code{sbm} object}

\item{kappa}{number of blocks in matrix}
}
\value{
matrix with \code{kappa} rows and a 1 at \code{(k,i)} if node \code{i} is in block \code{k} under \code{SBM}
}
\description{
converts block assignments of an \code{sbm} object to a matrix of block assignments
}
