% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocks.R
\name{blockmod}
\alias{blockmod}
\title{create a block-model object}
\usage{
blockmod(gamma, fixkappa, rblocks, logdblocks, dmarg)
}
\arguments{
\item{gamma}{parameters for the block model}

\item{fixkappa}{bool - is kappa fixed or can it vary}

\item{rblocks}{function(n, gamma) samples node memberships from the block model, takes gamma as a parameter}

\item{logdblocks}{log density for the number of blocks distribution}

\item{dmarg}{marginal density for the block assignments}
}
\value{
a bockmod object
}
\description{
create a block-model object
}
