% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocks.R
\name{blocks}
\alias{blocks}
\title{creat a blocks object}
\usage{
blocks(z, kappa)
}
\arguments{
\item{z}{vector of block labels}

\item{kappa}{maximum number of blocks}
}
\value{
a blocks object
}
\description{
creat a blocks object
}
\examples{
b <- blocks(c(1,1,2,3,4,4), 4)
}
