% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traces.R
\name{blocktrace}
\alias{blocktrace}
\title{plot a trace of the blocks from MCMC samples}
\usage{
blocktrace(postz)
}
\arguments{
\item{postz}{output from sampler}
}
\value{
ggplot object of node assignments against iteration number
}
\description{
plot a trace of the blocks from MCMC samples
}
