% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocks.R
\name{crp}
\alias{crp}
\title{Chinese Restaurant Process model}
\usage{
crp(gamma)
}
\arguments{
\item{gamma}{parameter for CRP(gam)}
}
\value{
a block model representing a CRP(gamma) distribution
}
\description{
Chinese Restaurant Process model
}
