% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocks.R
\name{dblocks.blocks}
\alias{dblocks.blocks}
\title{density for the block assignments}
\usage{
\method{dblocks}{blocks}(blocks, blockmod, ...)
}
\arguments{
\item{blocks}{a blocks object containing block assignments}

\item{blockmod}{a blockmod object containing a model for the block assignments}
}
\value{
The density for the block assignments in blocks under blockmod
}
\description{
density for the block assignments
}
