% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocks.R
\name{dblocks.sbm}
\alias{dblocks.sbm}
\title{density for the block assignments}
\usage{
\method{dblocks}{sbm}(sbm, mod)
}
\arguments{
\item{sbm}{an sbm model which contains the block state}

\item{mod}{a model for the block assignments}
}
\value{
The density for the block assignments in sbm under mod
}
\description{
density for the block assignments
}
