% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{ddirichlet}
\alias{ddirichlet}
\title{Density of X~dirichlet(gam)}
\usage{
ddirichlet(x, gam, log = FALSE)
}
\arguments{
\item{x}{random variable in the d-dimensional simplex}

\item{gam}{a length d probability vector}

\item{log}{return the log-probability instead?}
}
\value{
the density
}
\description{
Density of X~dirichlet(gam)
}
