% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edges.R
\name{dedges.numeric}
\alias{dedges.numeric}
\title{likelihood of edges}
\usage{
\method{dedges}{numeric}(edges, pmat, edgemod)
}
\arguments{
\item{edges}{an edges object}

\item{pmat}{a matrix of parameters}

\item{edgemod}{an edgemod object}
}
\value{
likelihood of edges under the edgemod using parameters in matrix pmat which has same size as edgse$E
}
\description{
likelihood of edges
}
