% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edges.R
\name{dedges.params}
\alias{dedges.params}
\title{density of edges}
\usage{
\method{dedges}{params}(edges, params, blocks, edgemod)
}
\arguments{
\item{edges}{an edges object}

\item{params}{an params object}

\item{blocks}{a blocks object}

\item{edgemod}{an edgemod object}
}
\value{
matrix same size as edges$E with density of each edge
}
\description{
density of edges
}
