% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edges.R
\name{dedges}
\alias{dedges}
\title{density of edges}
\usage{
dedges(edges, x, ...)
}
\arguments{
\item{edges}{an edges object}

\item{x}{an object}
}
\value{
matrix same size as edges$E with density of each edge
}
\description{
density of edges
}
