% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edges.R
\name{dedges.sbm}
\alias{dedges.sbm}
\title{density of edges}
\usage{
\method{dedges}{sbm}(edges, sbm, mod, ...)
}
\arguments{
\item{edges}{an edges object}

\item{sbm}{an sbm object}

\item{mod}{a model object}
}
\value{
matrix same size as edges$E with density of each edge
}
\description{
density of edges
}
