% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rjmcmc.R
\name{delblock}
\alias{delblock}
\title{delete a block move}
\usage{
delblock(Sbm, Edges, Mod, rho = 1)
}
\arguments{
\item{rho}{probability of choosing to add a block}

\item{sbm}{an sbm object}

\item{edges}{an edges object}

\item{mod}{a model object}
}
\value{
an updated sbm object
}
\description{
delete a block move
}
