% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocks.R
\name{dma}
\alias{dma}
\title{Dirichlet Multionoial Allocation model}
\usage{
dma(gamma, delta)
}
\arguments{
\item{gamma}{parameter for dirichlet component Dir(gam, gam, ...)}

\item{delta}{parameter for Poison(delta) - distribution on number of blocks}
}
\value{
a block model representing a dma(gamma, delta) distribution
}
\description{
Dirichlet Multionoial Allocation model
}
