% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{dparams.params}
\alias{dparams.params}
\title{calculate density of parameters}
\usage{
\method{dparams}{params}(Params, PM, log = TRUE, ...)
}
\arguments{
\item{PM}{a parammod object}

\item{log}{return log density?}

\item{...}{additional parameters for PM$df density funtcion}

\item{params}{a params object}
}
\value{
density of parameters in sbm under mod
}
\description{
calculate density of parameters
}
