% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{dparams}
\alias{dparams}
\title{calculate density of parameters}
\usage{
dparams(x, ...)
}
\arguments{
\item{x}{an object}
}
\value{
density of parameters under x
}
\description{
calculate density of parameters
}
