% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{dparams.sbm}
\alias{dparams.sbm}
\title{calculate density of parameters}
\usage{
\method{dparams}{sbm}(sbm, mod, ...)
}
\arguments{
\item{sbm}{an sbm object}

\item{mod}{a model object}
}
\value{
density of parameters in sbm under mod
}
\description{
calculate density of parameters
}
