% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp.R
\name{drawblock.dp}
\alias{drawblock.dp}
\title{draw blocks in Dirichlet process sampler}
\usage{
drawblock.dp(i, currsbm, Edges, Mod)
}
\arguments{
\item{i}{node to update}

\item{currsbm}{current SBM state}

\item{edges}{an edges object}

\item{mod}{a model object}
}
\value{
updated sbm
}
\description{
draw blocks in Dirichlet process sampler
}
