% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs.R
\name{drawblock.gibbs}
\alias{drawblock.gibbs}
\title{Gibbs step for drawing new block assignments}
\usage{
drawblock.gibbs(i, Sbm, Edges, Mod)
}
\arguments{
\item{i}{the node to reassign}

\item{Sbm}{an sbm object}

\item{Edges}{an edges object}

\item{Mod}{a model object}
}
\value{
updated sbm object
}
\description{
Gibbs step for drawing new block assignments
}
