% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp.R
\name{drawblocks.dp}
\alias{drawblocks.dp}
\title{draw blocks in Dirichlet process sampler}
\usage{
drawblocks.dp(currsbm, Edges, Mod)
}
\arguments{
\item{currsbm}{current SBM state}

\item{edges}{an edges object}

\item{mod}{a model object}
}
\value{
updated sbm
}
\description{
draw blocks in Dirichlet process sampler
}
