% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs.R
\name{drawblocks.gibbs}
\alias{drawblocks.gibbs}
\title{Gibbs step for drawing new block assignments}
\usage{
drawblocks.gibbs(Sbm, Edges, Mod)
}
\arguments{
\item{Sbm}{an sbm object}

\item{Edges}{an edges object}

\item{Mod}{a model object}
}
\value{
updated sbm object
}
\description{
Gibbs step for drawing new block assignments
}
