% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawparams.R
\name{drawparam.params}
\alias{drawparam.params}
\title{simulate parameter for the kth parameter}
\usage{
\method{drawparam}{params}(Params, Blocks, k, Parammod, sigma)
}
\arguments{
\item{Params}{current params object}

\item{Blocks}{current blocks object}

\item{k}{index of parameter to update}

\item{Parammod}{model object for the parameters}

\item{sigma}{parameter for rw}
}
\value{
list(proposed params, log-jacobian term)
}
\description{
simulate parameter for the kth parameter
}
