% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawparams.R
\name{drawparam.sbm}
\alias{drawparam.sbm}
\title{simulate parameter for the kth parameter}
\usage{
\method{drawparam}{sbm}(currsbm, k, Mod, sigma)
}
\arguments{
\item{currsbm}{current state of sbm under the MCMC chain}

\item{k}{index of parameter to update}

\item{Mod}{model object for the sbm}

\item{sigma}{parameter for drawparam}
}
\value{
list(proposed sbm, log-jacobian term)
}
\description{
simulate parameter for the kth parameter
}
