% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawparams.R
\name{drawparams}
\alias{drawparams}
\title{simulate parameters for the given model in a Metropolis scheme}
\usage{
drawparams(currsbm, Edges, Mod, sigma = 0.1)
}
\arguments{
\item{currsbm}{current state of sbm under the MCMC chain}

\item{Edges}{edge state data}

\item{Mod}{model object for the sbm}

\item{sigma}{parameter for drawparam}
}
\value{
updated sbm object
}
\description{
simulate parameters for the given model in a Metropolis scheme
}
