% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edges.R
\name{edgemod}
\alias{edgemod}
\title{Class for edge models}
\usage{
edgemod(rg, dg, ...)
}
\arguments{
\item{rg}{function to simulate edges}

\item{dg}{function to calculate likelihood of edges}
}
\value{
an edgemod object
}
\description{
Class for edge models
}
