% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edges.R
\name{edges}
\alias{edges}
\title{Class for edge data}
\usage{
edges(e, sym, loops, ...)
}
\arguments{
\item{e}{a matrix or array representing the raw edge-state data}

\item{sym}{is the network symetric? (is edge-state ji = edgestate ji)}

\item{loops}{does the network contain self-loops? (edges from node i to i)}
}
\value{
an edges object
}
\description{
Class for edge data
}
