% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edges.R
\name{edges_bern}
\alias{edges_bern}
\title{Make an edgemod model with bernoulli edge-states}
\usage{
edges_bern(...)
}
\arguments{
\item{...}{additional parameters to pass to rbinom}
}
\value{
an edgemod representing edgestates which are Bernoulli distributed
}
\description{
Make an edgemod model with bernoulli edge-states
}
