% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edges.R
\name{edges_nbin}
\alias{edges_nbin}
\title{Make an edgemod model with NegativeBinomial edge-states}
\usage{
edges_nbin(...)
}
\arguments{
\item{...}{additional parameters to pass to rnbinom}
}
\value{
an edgemod representing edgestates which are NegativeBinomial distributed
}
\description{
Make an edgemod model with NegativeBinomial edge-states
}
