% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edges.R
\name{edges_norm}
\alias{edges_norm}
\title{Make an edgemod model with Normal edge-states}
\usage{
edges_norm(...)
}
\arguments{
\item{...}{additional parameters to pass to rnorm}
}
\value{
an edgemod representing edgestates which are Gaussian distributed
}
\description{
Make an edgemod model with Normal edge-states
}
