% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edges.R
\name{edges_pois}
\alias{edges_pois}
\title{Make an edgemod model with Poisson edge-states}
\usage{
edges_pois(...)
}
\arguments{
\item{...}{additional parameters to pass to rpois}
}
\value{
an edgemod representing edgestates which are Poisson distributed
}
\description{
Make an edgemod model with Poisson edge-states
}
