% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edges.R
\name{edgesfromi.edges}
\alias{edgesfromi.edges}
\title{get the edges eminating from node i}
\usage{
\method{edgesfromi}{edges}(Edges, i, kappa = 1)
}
\arguments{
\item{Edges}{an edges object}

\item{i}{the node incident to the returned edges}

\item{kappa}{number of blocks}
}
\value{
an edges object containing only the edges incident to i
}
\description{
get the edges eminating from node i
}
