% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enron.R
\docType{data}
\name{Enron}
\alias{Enron}
\title{A data set of counts of emails between email addresses
This is a non-symmetric network.
Nodes represent email address.
The edge-state ij between two email addresses i and j is the number of emails sent from i to j
The Groups vector is the node label from the igraph attribute "notes"}
\format{A list containing
\describe{
  \item{Edges}{an edges object with each edge-state representing the number of emails between two email addresses}
  \item{Groups}{A vecto giving a group name to which the email address belong. The order matches the edges such that Edges[i,j] is the edge-state between the i-th and j-th nodes who are members of Groups[i] and Groups[j] respectively}
}}
\source{
https://cran.r-project.org/package=igraphdata
}
\usage{
Enron
}
\description{
A data set of counts of emails between email addresses
This is a non-symmetric network.
Nodes represent email address.
The edge-state ij between two email addresses i and j is the number of emails sent from i to j
The Groups vector is the node label from the igraph attribute "notes"
}
\keyword{datasets}
