% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traces.R
\name{eval_plots}
\alias{eval_plots}
\title{get a set of evaluation plots from MCMC samples}
\usage{
eval_plots(output, burnin, theta_index)
}
\arguments{
\item{output}{from sampler}

\item{burnin}{burn-in period (a vector of iteration numbers to subset outputs)}

\item{theta_index}{which set of thetas to plot?}
}
\value{
list of ggplot objects (with descriptive names)
}
\description{
get a set of evaluation plots from MCMC samples
}
