% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbm.R
\name{is.sbm}
\alias{is.sbm}
\title{is.sbm}
\usage{
is.sbm(x)
}
\arguments{
\item{x}{an R object}
}
\value{
Logical indicating if \code{x} is an \code{\link{sbm}} object
}
\description{
Logical check if an object is an \code{\link{sbm}} object
}
