% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edges.R
\name{loglike.params}
\alias{loglike.params}
\title{likelihood of edges}
\usage{
\method{loglike}{params}(edges, params, blocks, edgemod, na.rm = F)
}
\arguments{
\item{edges}{an edges object}

\item{params}{a params object}

\item{blocks}{a blocks object}

\item{edgemod}{an edgemod object}

\item{na.rm=F}{passed to sum}
}
\value{
likelihood of edges under the edgemod using parameters in matrix pmat which has same size as edgse$E
}
\description{
likelihood of edges
}
