% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edges.R
\name{loglike}
\alias{loglike}
\title{likelihood of edges}
\usage{
loglike(edges, x, ...)
}
\arguments{
\item{edges}{an edges object}

\item{x}{an object}
}
\value{
numeric likelihood of edges under x
}
\description{
likelihood of edges
}
