% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edges.R
\name{loglike.sbm}
\alias{loglike.sbm}
\title{likelihood of edges}
\usage{
\method{loglike}{sbm}(edges, sbm, mod, ...)
}
\arguments{
\item{edges}{an edges object}

\item{sbm}{an sbm object}

\item{mod}{a model object}
}
\value{
numeric likelihood of edges under x
}
\description{
likelihood of edges
}
