% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Macaque}
\alias{Macaque}
\title{The Macaque data set as extracted from \code{igraph} using the script in data-raw}
\format{
An \code{edges} object of activation counts between brain regions in a Macaque
}
\usage{
Macaque
}
\description{
The Macaque data set as extracted from \code{igraph} using the script in data-raw
}
\seealso{
igraph
}
\keyword{datasets}
