% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marglike.R
\name{marglike}
\alias{marglike}
\title{calculate the marginal likelihood for a node for samplers using conjugate models}
\usage{
marglike(znoi, ei, parammod, ...)
}
\arguments{
\item{ei}{edge-states incident to i}

\item{parammod}{a parammod object}

\item{znoti}{a matrix of blockassignments without node i}
}
\description{
calculate the marginal likelihood for a node for samplers using conjugate models
}
