% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marglike.R
\name{marglike_norm}
\alias{marglike_norm}
\title{Marginal likelihood model for Normal distributed edges}
\usage{
marglike_norm(znoi, ei, parammod)
}
\arguments{
\item{znoi}{a matrix of block assignments without node i}

\item{ei}{edge-states incident to i}

\item{parammod}{a \code{parammod} object}
}
\value{
log-probability of node i belonging to each block
}
\description{
calculate the marginal likelihood for a node for samplers using conjugate models
}
