% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocks.R
\name{margprior}
\alias{margprior}
\title{Calculate the marginal density for block assignment of node i under the model in BM and block assignments in blocks}
\usage{
margprior(blocks, BM, i, ...)
}
\arguments{
\item{blocks}{blocks object}

\item{BM}{a blockmod object}

\item{i}{a node index}
}
\value{
marginal density for block assignment of node i under the model in BM and block assignments in blocks
}
\description{
Calculate the marginal density for block assignment of node i under the model in BM and block assignments in blocks
}
