% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rjmcmc.R
\name{mergeavg}
\alias{mergeavg}
\title{merge move using average to merge parameters}
\usage{
mergeavg(sbm, edges, mod, ...)
}
\arguments{
\item{sbm}{an sbm object}

\item{edges}{an edges object}

\item{mod}{a model object}

\item{...}{additional parameter to 'accept'}
}
\value{
an updated sbm object
}
\description{
merge move using average to merge parameters
}
