% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rjmcmc.R
\name{mergeblocks}
\alias{mergeblocks}
\title{merge move block merging}
\usage{
mergeblocks(sbm, k, l, edges, mod, ...)
}
\arguments{
\item{sbm}{an sbm object}

\item{k, l}{blocks to merge}

\item{edges}{an edges object}

\item{mod}{a model object}

\item{...}{additional parameter to 'accept'}
}
\value{
list(proposed block structure, log-acceptance-prob)
}
\description{
merge move block merging
}
