% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rjmcmc.R
\name{mergeparams.default}
\alias{mergeparams.default}
\title{Merge step - parameter merging}
\usage{
\method{mergeparams}{default}(params, k, l, pm)
}
\arguments{
\item{params}{a params object}

\item{k, l}{blocks to merge}

\item{pm}{a parammod object}
}
\value{
list(proposed_params, log-acceptance-prob)
}
\description{
Merge step - parameter merging
}
