% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rjmcmc.R
\name{mergeparams.numeric}
\alias{mergeparams.numeric}
\title{Merge step - parameter merging}
\usage{
\method{mergeparams}{numeric}(thetak, thetal, x, pm)
}
\arguments{
\item{thetak, thetal}{parameters to merge}

\item{x}{auxiliary parameter}

\item{pm}{a parammod object}
}
\value{
list(proposed_params, log-acceptance-prob)
}
\description{
Merge step - parameter merging
}
