% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rjmcmc.R
\name{mergeparams}
\alias{mergeparams}
\title{merge parameters}
\usage{
mergeparams(x, ...)
}
\arguments{
\item{x}{an object to dispatch on}

\item{...}{additional arguments for methods}
}
\value{
merged parameters from \code{x}
}
\description{
merge parameters
}
\seealso{
\code{\link{mergeparams.default}} \code{\link{mergeparams.numeric}}
}
