% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traces.R
\name{modeblocks}
\alias{modeblocks}
\title{modal block assignments from MCMC samples}
\usage{
modeblocks(postz)
}
\arguments{
\item{postz}{output from sampler}
}
\value{
a blocks object with the modal block assignments under \code{postz}
}
\description{
modal block assignments from MCMC samples
}
