% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocks.R
\name{multinom}
\alias{multinom}
\title{Multionoial block assignment}
\usage{
multinom(gamma)
}
\arguments{
\item{gamma}{parameter for dirichlet component Dir(gam, gam, .., gam)}
}
\value{
a block model representing a multinomial(gamma) distribution
}
\description{
Multionoial block assignment
}
