% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edges.R
\name{nodelike.blocks}
\alias{nodelike.blocks}
\title{get likelihood of edges eminating from node i}
\usage{
\method{nodelike}{blocks}(Edges, Blocks, Params, Mod, i, ...)
}
\arguments{
\item{Edges}{an edges object}

\item{Params}{a params object}

\item{Mod}{a model object}

\item{i}{the node of interest}

\item{blocks}{an blocks object}
}
\value{
likelihood of edges eminating from node i
}
\description{
get likelihood of edges eminating from node i
}
