% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edges.R
\name{nodelike}
\alias{nodelike}
\title{get likelihood of edges eminating from node i}
\usage{
nodelike(edges, x, ...)
}
\arguments{
\item{x}{an object}

\item{Edges}{an edges object}
}
\value{
likelihood of edges eminating from node i
}
\description{
get likelihood of edges eminating from node i
}
