% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edges.R
\name{nodelike.sbm}
\alias{nodelike.sbm}
\title{get likelihood of edges eminating from node i}
\usage{
\method{nodelike}{sbm}(Edges, Sbm, Mod, i, ...)
}
\arguments{
\item{Edges}{an edges object}

\item{Mod}{a model object}

\item{i}{the node of interest}

\item{sbm}{an sbm object}
}
\value{
likelihood of edges eminating from node i
}
\description{
get likelihood of edges eminating from node i
}
