% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traces.R
\name{numblockstrace}
\alias{numblockstrace}
\title{plot a trace of the number of blocks from MCMC samples}
\usage{
numblockstrace(postk)
}
\arguments{
\item{postk}{output from sampler}
}
\value{
ggplot object of kappa against iteration number
}
\description{
plot a trace of the number of blocks from MCMC samples
}
