% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{param_beta}
\alias{param_beta}
\title{beta parameter model}
\usage{
param_beta(a0, a1, b0, b1)
}
\arguments{
\item{a0}{theta_0 ~ Beta(a0,a1)}

\item{a1}{theta_0 ~ Beta(a0,a1)}

\item{b0}{theta_k ~ Beta(b0,b1)}

\item{b1}{theta_k ~ Beta(b0,b1)}
}
\value{
parammod representing beta distributed parameters
}
\description{
beta parameter model
}
