% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{param_gamma}
\alias{param_gamma}
\title{gamma parameter model}
\usage{
param_gamma(a0, a1, b0, b1)
}
\arguments{
\item{a0}{theta_0 ~ Gamma(a0,a1)}

\item{a1}{theta_0 ~ Gamma(a0,a1)}

\item{b0}{theta_k ~ Gamma(b0,b1)}

\item{b1}{theta_k ~ Gamma(b0,b1)}
}
\value{
parammod representing gamma distributed parameters
}
\description{
gamma parameter model
}
