% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{param_nbin}
\alias{param_nbin}
\title{Negative Binomial parameter model
theta_0 = (mu0, sigma0)
theta_k = (muk, sigmak)}
\usage{
param_nbin(a0, a1, b0, b1, c0, c1, d0, d1)
}
\arguments{
\item{a0, a1}{mu0 ~ Gamma(a0,a1)}

\item{b0, b1}{sig_0 ~ Beta(b0,b1)}

\item{c0, c1}{muk ~ Gamma(c0, c1)}

\item{d0, d1}{sig_k ~ Beta(d0,d1)}
}
\value{
parammod representing NegativeBinomial distributed parameters
}
\description{
Negative Binomial parameter model
theta_0 = (mu0, sigma0)
theta_k = (muk, sigmak)
}
