% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{param_norm}
\alias{param_norm}
\title{Normal parameter model
theta_0 = (mu0, sigma0)
theta_k = (muk, sigmak)}
\usage{
param_norm(a0, a1, b0, b1, c0, c1, d0, d1)
}
\arguments{
\item{a0, a1}{mu0 ~ Normal(a0,a1)}

\item{b0, b1}{sig_0 ~ Gamma(b0,b1)}

\item{c0, c1}{muk ~ Normal(c0, c1)}

\item{d0, d1}{sig_k ~ Gamma(d0,d1)}
}
\value{
parammod representing Normal distributed parameters
}
\description{
Normal parameter model
theta_0 = (mu0, sigma0)
theta_k = (muk, sigmak)
}
