% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{parammat.blocks}
\alias{parammat.blocks}
\title{Helper function -  Make a matrix of parameters}
\usage{
\method{parammat}{blocks}(blocks, params, ...)
}
\arguments{
\item{blocks}{a blocks object}

\item{params}{the parameters object}
}
\value{
a matrix of parameters of size NxN where parammet(i,j) is the parameter governing edge i,j under the block assignments in blocks
}
\description{
Helper function -  Make a matrix of parameters
}
