% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{parammat.matrix}
\alias{parammat.matrix}
\title{Helper function -  Make a matrix of parameters}
\usage{
\method{parammat}{matrix}(zleft, zright, params, ...)
}
\arguments{
\item{zleft}{block assignments on the left}

\item{zright}{block assignments on the right}

\item{params}{the parameters object}
}
\value{
a matrix of parameters of size |zleft| x |zright|
}
\description{
Helper function -  Make a matrix of parameters
}
