% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{parammat.params}
\alias{parammat.params}
\title{Helper function -  Make a matrix of parameters}
\usage{
\method{parammat}{params}(Params, kappa)
}
\arguments{
\item{kappa}{- number of blocks to compute for - needed in computing the sequential assignmetn during split move}

\item{params}{the parameters object}
}
\value{
a matrix of parameters
}
\description{
Helper function -  Make a matrix of parameters
}
