% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parammat.R
\name{parammat}
\alias{parammat}
\title{Parameter Matrix}
\usage{
parammat(x, ...)
}
\arguments{
\item{x}{object for dispatch}

\item{...}{additional arguments for method}
}
\value{
a parameter matrix object
}
\description{
Make a matrix of parameters
}
