% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{parammat.sbm}
\alias{parammat.sbm}
\title{Helper function -  Make a matrix of parameters}
\usage{
\method{parammat}{sbm}(sbm)
}
\arguments{
\item{sbm}{an SBM object}
}
\description{
Helper function -  Make a matrix of parameters
}
