% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{parammod}
\alias{parammod}
\title{Class parammod}
\usage{
parammod(alpha, beta, rf, df, t, invt, loggradt)
}
\arguments{
\item{alpha}{parameters for theta_0}

\item{beta}{parameters for theta_k}

\item{rf}{random function to draw parameters (the prior) rf(n, x) draws n parameters with hyperparameters x}

\item{df}{density function for parameters df(theta, x) density of theta with hyperparameters x}

\item{t}{map taking parameter space to real line}

\item{invt}{map taking real line to parameter space}

\item{loggradt}{log of the gradient of the map t}
}
\value{
a parammod object
}
\description{
Class parammod
}
