% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{params}
\alias{params}
\title{Class params}
\usage{
params(theta0, thetak)
}
\arguments{
\item{theta0}{between block parameter}

\item{thetak}{within block parameters}
}
\value{
a params object
}
\description{
Class params
}
