% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traces.R
\name{paramtrace}
\alias{paramtrace}
\title{plot a trace of parameter values from MCMC samples}
\usage{
paramtrace(theta, truetheta)
}
\arguments{
\item{theta}{output from sampler}

\item{truetheta}{if supplied, adds guidelines for the true parameters}
}
\value{
ggplot object of theta value against iteration number
}
\description{
plot a trace of parameter values from MCMC samples
}
