% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocks.R
\name{plot.blocks}
\alias{plot.blocks}
\title{plots a block object}
\usage{
\method{plot}{blocks}(blocks, col, xaxt = "n", yaxt = "n", xlab = "",
  ylab = "", ...)
}
\arguments{
\item{blocks}{a blocks object to plot}

\item{col}{colors for the plot}

\item{...}{additional parameters for 'image'}

\item{xaxt='n'}{override 'image' parameters}

\item{yaxt='n'}{override 'image' parameters}

\item{xlab=""}{override 'image' parameters}

\item{ylab=""}{override 'image' parameters}
}
\description{
plots a block object
}
