% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edges.R
\name{plot.edges}
\alias{plot.edges}
\title{plots edges objects}
\usage{
\method{plot}{edges}(edges, blocks, sorted = TRUE, xlab = "Node",
  ylab = " ", ...)
}
\arguments{
\item{edges}{an edges object}

\item{blocks}{a blocks object or sbm object}

\item{sorted}{sort by block membership in blocks before plotting?}

\item{...}{parameters for image}
}
\description{
plots edges objects
}
