% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbm.R
\name{plot.sbm}
\alias{plot.sbm}
\alias{image.sbm}
\title{Plot for \code{\link{sbm}} object}
\usage{
\method{plot}{sbm}(x, col, ...)

\method{image}{sbm}(x, col, ...)
}
\arguments{
\item{x}{an \code{\link{sbm}} object}

\item{col}{colours for each block - if missing, \code{rainbow} is used}

\item{...}{additional arguments for plot}
}
\description{
plot an \code{\link{sbm}} object as an \code{image}
}
\seealso{
plot.default
}
