% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traces.R
\name{plotpostpairs}
\alias{plotpostpairs}
\title{helper function for trace plots}
\usage{
plotpostpairs(mat)
}
\arguments{
\item{mat}{matrix to plot as an image using ggplot2}
}
\value{
`ggplot2` plot objecy
}
\description{
helper function for trace plots
}
