% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traces.R
\name{postpairs}
\alias{postpairs}
\title{mean proportion of times two nodes were in the same block under MCMC samples}
\usage{
postpairs(postz)
}
\arguments{
\item{postz}{output from sampler}
}
\value{
matrix P with P[i,j] = proportion of times i and j are in the same block under \code{postz}
}
\description{
mean proportion of times two nodes were in the same block under MCMC samples
}
