% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocks.R
\name{rblocks}
\alias{rblocks}
\title{random block assignment generation}
\usage{
rblocks(n, BM, sorted = FALSE, ...)
}
\arguments{
\item{n}{number of nodes}

\item{BM}{a blockmod object}

\item{sorted}{sort the nodes by block assignment for nice plots?}

\item{...}{additional parameters for the BM$rblocks function}
}
\value{
a blocks object
}
\description{
random block assignment generation
}
