% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{rcat}
\alias{rcat}
\title{Random draw from Categorical(p)}
\usage{
rcat(n, p, replace = TRUE)
}
\arguments{
\item{n}{number of draws}

\item{p}{a length-d probability vector}

\item{replace}{should the categories be replaced? If so n < p required}
}
\value{
a length-d vector x, where \eqn{x[i] \in \{1,...,d\}}
}
\description{
Random draw from Categorical(p)
}
