% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{rdirichlet}
\alias{rdirichlet}
\title{Draw n random variables from Dricihlet(gam) - gam a vector of length K}
\usage{
rdirichlet(n, gam)
}
\value{
x = Matrix(n,K)
}
\description{
Draw n random variables from Dricihlet(gam) - gam a vector of length K
}
