% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edges.R
\name{redges.params}
\alias{redges.params}
\title{simulate edges}
\usage{
\method{redges}{params}(params, blocks, edgemod, sym = T, loops = F,
  ...)
}
\arguments{
\item{params}{a params object}

\item{blocks}{a blocks object}

\item{edgemod}{an edgemod object}

\item{sym}{should the network be symmetric?}

\item{loops}{should the network have self-loops?}
}
\value{
an edges object
}
\description{
simulate edges
}
