% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edges.R
\name{redges.sbm}
\alias{redges.sbm}
\title{simulate edges}
\usage{
\method{redges}{sbm}(sbm, mod, ...)
}
\arguments{
\item{sbm}{an sbm object}

\item{mod}{a model object}
}
\value{
an edges object
}
\description{
simulate edges
}
