% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rjmcmc.R
\name{rmblock}
\alias{rmblock}
\title{remove a block}
\usage{
rmblock(z, k)
}
\arguments{
\item{z}{factor vector of block memberships}

\item{k}{a block (level of z)}
}
\value{
blocks object with no block k
}
\description{
remove a block
}
