% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawparams.R
\name{rparam}
\alias{rparam}
\title{draw from the prior - locjac = 0}
\usage{
rparam(k, pm)
}
\arguments{
\item{k}{index of parameter}

\item{pm}{parammodel - contains the transform/map to project parameter to the real line and the prior on parameters}
}
\description{
draw from the prior - locjac = 0
}
