% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{rparams}
\alias{rparams}
\title{generate parameters}
\usage{
rparams(kappa, PM)
}
\arguments{
\item{kappa}{number of blocks}

\item{PM}{a parammod object}
}
\value{
a params object
}
\description{
generate parameters
}
