% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawparams.R
\name{rw}
\alias{rw}
\title{random walk}
\usage{
rw(p, pm, sigma)
}
\arguments{
\item{p}{a parameter}

\item{pm}{parammodel - contains the transform/map to project parameter to the real line}

\item{sigma}{- scale of random walk}
}
\description{
random walk
}
