% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampler.R
\name{sampler.conj}
\alias{sampler.conj}
\title{Conjugate model sampler}
\usage{
sampler.conj(currsbm, edges, sbmmod, ...)
}
\arguments{
\item{currsbm}{the current state of the sampler}

\item{edges}{an \code{\link{edges}} object}

\item{sbmmod}{an \code{\link{sbmmod}} model}

\item{...}{additional arguments for \code{sbmmod$marglike}}
}
\value{
next state of \code{currsbm} object
}
\description{
Conjugate model sampler
}
\examples{
model <- sbmmod(multinom(2, 3), param_beta(1,1,1,1), edges_bern(), marglike=marglike_bern)
trueSBM <- model$r(100)
Edges <- redges(trueSBM, model$edge)
\dontrun{out <- sampler(Edges, model, 100, "conjugate")}
}
