% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampler.R
\name{sampler.gibbs}
\alias{sampler.gibbs}
\title{Gibbs sampling for node assignments}
\usage{
sampler.gibbs(currsbm, edges, sbmmod, sigma)
}
\arguments{
\item{currsbm}{the current state of the sampler}

\item{edges}{an \code{\link{edges}} object}

\item{sbmmod}{an \code{\link{sbmmod}} model}

\item{sigma}{random walk parameter for theta}
}
\value{
next state of \code{currsbm} object
}
\description{
Gibbs sampling for node assignments
}
\examples{
model <- sbmmod(multinom(1, 3), param_beta(1,1,1,1), edges_bern())
trueSBM <- model$r(100)
Edges <- redges(trueSBM, model$edge)
\dontrun{gibbs_out <- sampler(Edges, model, algorithm="gibbs", 100, sigma=0.1)}
\dontrun{eval_plots(gibbs_out)}
}
