% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampler.R
\name{sampler}
\alias{sampler}
\title{top level sampler function}
\usage{
sampler(Edges, Mod, nSteps = 1000, algorithm = "rj", sigma = 0.5,
  statusfreq = 100, currsbm = NULL, ...)
}
\arguments{
\item{nSteps}{number of steps to run mcmc for}

\item{algorithm}{choice of algorithm options are: "conjugate", "gibbs", "dp", "rj"}

\item{sigma}{random walk parameter for theta}

\item{statusfreq}{frequency to print some status info out - in number of iterations}

\item{currsbm}{initial state for sbm object - if NULL one is drawn from mof}

\item{...}{additional parameters to pass to sbm for currsbm}

\item{edges}{edges object}

\item{mod}{model object}
}
\value{
postz traces for block assignments z

postt traces for theta

postk traces for number of blocks kappa

postn traces for number of occupied blocks

nsteps how many steps did the algo use

algorithm which algo was used
}
\description{
top level sampler function
}
