% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbm.R
\name{sbm}
\alias{sbm}
\title{Class \code{sbm}}
\usage{
sbm(blocks, params)
}
\arguments{
\item{blocks}{a \code{blocks} object}

\item{params}{a \code{params} object}
}
\value{
an \code{sbm} object
}
\description{
Class \code{sbm}
}
\examples{
sbm(blocks(c(1,1,2,2,3,3)), params(0.1, c(0.4,0.5,0.6)))
}
