% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbmmod.R
\name{sbmmod}
\alias{sbmmod}
\title{Stochastic block model object}
\usage{
sbmmod(blockmod, parammod, edgemod, ...)
}
\arguments{
\item{blockmod}{a \code{\link{blockmod}} object}

\item{parammod}{a \code{\link{parammod}} object}

\item{edgemod}{an \code{\link{edgemod}} object}

\item{...}{additional arguments to store in the \code{sbmmod} object}
}
\value{
an \code{sbmmod} object with a method `r(n)` sampling an \code{sbm} object with \code{n} nodes from the model and a method \code{logd(sbm)} computing the log-density of \code{sbm} under the model
}
\description{
A wrapper for a block and parameter model
}
\details{
Simple wrapper for the block and parameter model for an \code{sbm} object
}
\seealso{
\code{\link{blockmod}} \code{\link{parammod}} \code{\link{edgemod}}
}
\author{
Matthew Ludkin
}
