% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rjmcmc.R
\name{splitavg}
\alias{splitavg}
\title{split move using average to merge parameters}
\usage{
splitavg(Sbm, Edges, Mod, ...)
}
\arguments{
\item{...}{additional parameter to 'accept'}

\item{sbm}{an sbm object}

\item{edges}{an edges object}

\item{mod}{a model object}
}
\value{
an updated sbm object
}
\description{
split move using average to merge parameters
}
