% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rjmcmc.R
\name{splitblocks}
\alias{splitblocks}
\title{split move block mergingsplitting}
\usage{
splitblocks(Blocks, pparams, k, Edges, Mod, ...)
}
\arguments{
\item{Blocks}{a blocks object}

\item{pparams}{propoed_params}

\item{k}{block to split}

\item{Edges}{an edges object}

\item{...}{additional parameter to 'accept'}

\item{mod}{a model object}
}
\value{
list(proposed block structure, log-acceptance-prob)
}
\description{
split move block mergingsplitting
}
