% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rjmcmc.R
\name{splitparams.numeric}
\alias{splitparams.numeric}
\title{split move: \code{params}}
\usage{
\method{splitparams}{numeric}(theta, u, x, parammod)
}
\arguments{
\item{theta}{a parameter to split}

\item{u}{auxiliary variable}

\item{x}{auxiliary variable}

\item{parammod}{\code{parammod} object}
}
\value{
\code{list(proposed_params, log-acceptance-prob)}
}
\description{
split move: \code{params}
}
