% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rjmcmc.R
\name{splitparams.params}
\alias{splitparams.params}
\title{split move split params}
\usage{
\method{splitparams}{params}(params, k, pm)
}
\arguments{
\item{params}{a parameter object to split}

\item{k}{block to split}

\item{pm}{parammod object}
}
\value{
list(proposed_params, log-acceptance-prob)
}
\description{
split move split params
}
