% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{StackOverflow}
\alias{StackOverflow}
\title{The Stack-Overflow data set as extracted from \code{igraph} using the script in data-raw
Extracted on 27/8/2019 from Kaggle (login required) using:
\code{library(rvest)}
\code{read_html("https://www.kaggle.com/stackoverflow/stack-overflow-tag-network/downloads/stack_network_links.csv/1")}}
\format{
An \code{edges} object of activation counts between brain regions in a Macaque
}
\source{
\url{https://www.kaggle.com/stackoverflow/stack-overflow-tag-network/}
}
\usage{
StackOverflow
}
\description{
The Stack-Overflow data set as extracted from \code{igraph} using the script in data-raw
Extracted on 27/8/2019 from Kaggle (login required) using:
\code{library(rvest)}
\code{read_html("https://www.kaggle.com/stackoverflow/stack-overflow-tag-network/downloads/stack_network_links.csv/1")}
}
\seealso{
igraph
}
\keyword{datasets}
