% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocks.R
\name{updateblock.blocks}
\alias{updateblock.blocks}
\title{Update the block assignment of a node}
\usage{
\method{updateblock}{blocks}(blocks, newblock, i)
}
\arguments{
\item{blocks}{a blocks object}

\item{newblock}{the new block for node i}

\item{i}{the node to update}
}
\value{
new blocks object
}
\description{
Update the block assignment of a node
}
