% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateblock.R
\name{updateblock}
\alias{updateblock}
\title{Update the block assignment of a node}
\usage{
updateblock(x, ...)
}
\arguments{
\item{x}{object for dispatch}

\item{...}{additional arguments for method}
}
\value{
object like `x` with updated block structure
}
\description{
change the block assignment in \code{x} of a node to a new block
}
\seealso{
\code{\link{updateblock.blocks}} \code{\link{updateblock.sbm}}
}
