% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateblock.R
\name{updateblock.sbm}
\alias{updateblock.sbm}
\title{Update the block assignment of a node}
\usage{
\method{updateblock}{sbm}(sbm, i, newblock)
}
\arguments{
\item{sbm}{an \code{\link{sbm}} object}

\item{i}{the node to update}

\item{newblock}{the new block for node i}
}
\value{
new \code{sbm} object
}
\description{
change the block assignment in an \code{sbm} object to a new block
}
