% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{vmeasure}
\alias{vmeasure}
\title{V-measure}
\usage{
vmeasure(z, truez, beta = 1)
}
\arguments{
\item{z}{input vector}

\item{truez}{reference vector}

\item{beta}{vmeasure parameter beta=1 gives equal weight to homogeneity and completeness}
}
\value{
v-measure of z against truez
}
\description{
V-measure
}
