% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocks.R
\name{zmat.blocks}
\alias{zmat.blocks}
\title{converts blocks to a matrix of block assignments}
\usage{
\method{zmat}{blocks}(blocks, kappa)
}
\arguments{
\item{blocks}{a blocks object}

\item{kappa}{number of blocks}
}
\value{
matrix with K rows and a 1 at (k,i) if node i is in block k under the blocks object
}
\description{
converts blocks to a matrix of block assignments
}
