% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocks.R
\name{zmat.numeric}
\alias{zmat.numeric}
\title{converts x to a matrix of block assignments}
\usage{
\method{zmat}{numeric}(x, K)
}
\arguments{
\item{x}{a vector of node-to-block assignments}

\item{K}{number of blocks}
}
\value{
matrix with K rows and a 1 at (k,i) if node i is in block k under x
}
\description{
converts x to a matrix of block assignments
}
