% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocks.R
\name{zmat.sbm}
\alias{zmat.sbm}
\title{converts the block assignment under sbm to a matrix of block assignments}
\usage{
\method{zmat}{sbm}(sbm)
}
\arguments{
\item{sbm}{an sbm object}
}
\value{
matrix with K rows and a 1 at (k,i) if node i is in block k under the sbm model
}
\description{
converts the block assignment under sbm to a matrix of block assignments
}
