% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampler.R
\name{accept}
\alias{accept}
\title{accept \code{propsbm} with the acceptance probability alpha}
\usage{
accept(currsbm, propsbm, edges, sbmmod, logjac = 0, logu = 0, ...)
}
\arguments{
\item{currsbm}{current \code{sbm} state}

\item{propsbm}{proposed \code{sbm} state}

\item{edges}{an \code{\link{edges}} object}

\item{sbmmod}{an \code{\link{sbmmod}} model}

\item{logjac}{log Jacobian of transformation of variables}

\item{logu}{log density for auxiliary variables}

\item{...}{additional arguments to pass to \code{dedges}}
}
\value{
updated \code{sbm} object
}
\description{
accept \code{propsbm} with the acceptance probability alpha
}
