% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rjmcmc.R
\name{addblock}
\alias{addblock}
\title{Add a block move}
\usage{
addblock(sbm, edges, sbmmod, rho = 1)
}
\arguments{
\item{sbm}{the current state of the sampler}

\item{edges}{an \code{\link{edges}} object}

\item{sbmmod}{an \code{\link{sbmmod}} model}

\item{rho}{probability of choosing to add a block}
}
\value{
an updated \code{sbm} object
}
\description{
proposes adding an empty block labelled \code{kappa+}1 to \code{sbm}
}
