% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocks.R
\name{blocks}
\alias{blocks}
\title{Blocks object}
\usage{
blocks(z, kappa)
}
\arguments{
\item{z}{vector of block labels for each node}

\item{kappa}{maximum number of blocks}
}
\value{
a \code{blocks} object
}
\description{
create a blocks object
}
\details{
stores the block allocations and total number of blocks for a stochastic block model
}
\examples{
## Assign six nodes to four blocks:
b <- blocks(c(1,1,2,3,4,4), 4)
print(b)
plot(b) ## shows id two nodes are members of the same block
}
