% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{ddirichlet}
\alias{ddirichlet}
\title{Dirichlet distribution}
\usage{
ddirichlet(x, gam, log = FALSE)
}
\arguments{
\item{x}{random variable in the d-dimensional simplex}

\item{gam}{a length K concentration parameter}

\item{log}{return the log-probability instead?}
}
\value{
the density
}
\description{
Density of Dirichlet distribution
}
\examples{
g <- rep(2,5)
p <- rdirichlet(1, g) ## a length-5 probability vector
ddirichlet(p, g)
}
