% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edges.R
\name{dedges.numeric}
\alias{dedges.numeric}
\title{likelihood of edges}
\usage{
\method{dedges}{numeric}(x, edges, edgemod, na.rm = na.rm, ...)
}
\arguments{
\item{x}{a matrix of parameters (with same size as \code{edges$E})}

\item{edges}{an \code{\link{edges}} object}

\item{edgemod}{an \code{\link{edgemod}} object}

\item{na.rm}{remove NAs when calculating?}

\item{...}{additional arguments passed to \code{edgemod$logd}}
}
\value{
likelihood of edges under the \code{\link{edgemod}} using parameters in matrix \code{pmat}
}
\description{
likelihood of edges
}
