% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edges.R
\name{dedges}
\alias{dedges}
\title{Density of edges}
\usage{
dedges(x, edges, edgemod, na.rm = TRUE, ...)
}
\arguments{
\item{x}{an R object for dispatch}

\item{edges}{an \code{\link{edges}} object}

\item{edgemod}{an \code{\link{edgemod}} object}

\item{na.rm}{remove NAs when calculating?}

\item{...}{additional arguments}
}
\value{
matrix same size as \code{edges$E} with density of each edge
}
\description{
Compute the probability density for an \code{\link{edges}} object
}
\seealso{
\code{\link{dedges.sbm}} \code{\link{dedges.sbm}}
}
