% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rjmcmc.R
\name{delblock}
\alias{delblock}
\title{Delete a block move}
\usage{
delblock(sbm, edges, sbmmod, rho = 1)
}
\arguments{
\item{sbm}{the current state of the sampler}

\item{edges}{an \code{\link{edges}} object}

\item{sbmmod}{an \code{\link{sbmmod}} model}

\item{rho}{probability of choosing to add a block}
}
\value{
an updated \code{sbm} object
}
\description{
proposes deleting an empty block (chosen at random among empty Blocks)
}
