% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp.R
\name{drawblock.dp}
\alias{drawblock.dp}
\title{Draw block membership}
\usage{
drawblock.dp(i, currsbm, edges, sbmmod)
}
\arguments{
\item{i}{node to update}

\item{currsbm}{current \code{sbm} object}

\item{edges}{an \code{\link{edges}} object}

\item{sbmmod}{an \code{\link{sbmmod}} object}
}
\value{
updated \code{sbm} object
}
\description{
Draw block membership in a Dirichlet process sampler
}
\details{
sample a new block assignment for i under a Dirichlet process.
Care needs to be taken with singleton blocks to update the parameter model in \code{currsbm}.
}
\seealso{
For full algorithm details see \url{http://doi.org/10.17635/lancaster/thesis/296}
}
