% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs.R
\name{drawblock.gibbs}
\alias{drawblock.gibbs}
\title{Gibbs-like reassignment of nodes to the current set of blocks}
\usage{
drawblock.gibbs(i, currsbm, edges, sbmmod)
}
\arguments{
\item{i}{the node to reassign}

\item{currsbm}{an \code{\link{sbm}} object}

\item{edges}{an \code{\link{edges}} object}

\item{sbmmod}{an \code{\link{sbmmod}} object}
}
\value{
updated \code{sbm} object with new block assignment for i
}
\description{
Reassign node `i`  to the current set of blocks given the current number of blocks and the other block assignments
}
