% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp.R
\name{drawblocks.dp}
\alias{drawblocks.dp}
\title{Draw block memberships}
\usage{
drawblocks.dp(currsbm, edges, sbmmod)
}
\arguments{
\item{currsbm}{current \code{\link{sbm}} object}

\item{edges}{an \code{\link{edges}} object}

\item{sbmmod}{an \code{\link{sbmmod}} object}
}
\value{
updated \code{sbm} object
}
\description{
Draw block memberships in a Dirichlet process sampler
}
\details{
iteratively updates the block  assignment of each node using a Dirichlet process update move
}
