% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs.R
\name{drawblocks.gibbs}
\alias{drawblocks.gibbs}
\title{Gibbs-like reassignment of nodes to the current set of blocks}
\usage{
drawblocks.gibbs(currsbm, edges, sbmmod)
}
\arguments{
\item{currsbm}{an \code{\link{sbm}} object}

\item{edges}{an \code{\link{edges}} object}

\item{sbmmod}{an \code{\link{sbmmod}} object}
}
\value{
updated \code{\link{sbm}} object with new block assignments
}
\description{
Sweep through the set of nodes and reassign to the current set of blocks given the current number of blocks
}
