% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawparams.R
\name{drawparams}
\alias{drawparams}
\title{Metropolis updates by drawing parameters}
\usage{
drawparams(sbm, edges, sbmmod, sigma = 0.1)
}
\arguments{
\item{sbm}{current \code{\link{sbm}} object}

\item{edges}{an \code{\link{edges}}}

\item{sbmmod}{an \code{\link{sbmmod}}}

\item{sigma}{parameter for \code{drawparam}}
}
\value{
updated \code{sbm} object
}
\description{
Simulate parameters for the given model with a Metropolis-Hastings step
}
\details{
iterate through the parameters in \code{currsbm} and update.
}
