% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgemod.R
\name{edgemod}
\alias{edgemod}
\title{Class for edge models}
\usage{
edgemod(logd, r, ...)
}
\arguments{
\item{logd}{function(e, p) to calculate likelihood of edge an edge e given parameter array p}

\item{r}{function(p) - simulate an edge given a parameter p (optional)}

\item{...}{additional arguments to append to \code{edgemod} internal list}
}
\value{
an \code{edgemod} object
}
\description{
A class with a random and density method for \code{edges} objects
}
\note{
the parameter for \code{logd} is an array of c(dimension of theta, dim(E)) e.g. from \code{\link{parammat}}
}
\seealso{
\code{\link{edges_bern}} \code{\link{edges_pois}} \code{\link{edges_norm}}
}
