% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgemod.R
\name{edges_bern}
\alias{edges_bern}
\title{Bernoulli edge model}
\usage{
edges_bern(...)
}
\arguments{
\item{...}{additional parameters to pass to \code{rbinom}}
}
\value{
an \code{edgemod}
}
\description{
Make an \code{edgemod} model with Bernoulli edge-states
}
\examples{
eb <- edges_bern() ## makes `eb` an edgemod for Bernoulli edge-states
}
