% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgemod.R
\name{edges_nbin}
\alias{edges_nbin}
\title{Negative-Binomial edge model}
\usage{
edges_nbin(...)
}
\arguments{
\item{...}{additional parameters to pass to \code{rnbinom}}
}
\value{
an \code{edgemod}
}
\description{
Make an \code{edgemod} model with Negative-Binomial edge-states
}
\examples{
enb <- edges_nbin() ## makes `enb` an edgemod for Negative-Binomial edge-states
}
