% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgemod.R
\name{edges_norm}
\alias{edges_norm}
\title{Normal edge model}
\usage{
edges_norm(...)
}
\arguments{
\item{...}{additional parameters to pass to \code{rnorm}}
}
\value{
an \code{edgemod}
}
\description{
Make an \code{edgemod} model with Normal edge-states
}
\examples{
en <- edges_norm() ## makes `en` an edgemod for Normal edge-states
}
