% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgemod.R
\name{edges_pois}
\alias{edges_pois}
\title{Poisson edge model}
\usage{
edges_pois(...)
}
\arguments{
\item{...}{additional parameters to pass to \code{rpois}}
}
\value{
an \code{edgemod}
}
\description{
Make an \code{edgemod} model with Poisson edge-states
}
\examples{
ep <- edges_pois() ## makes `ep` an edgemod for Poisson edge-states
}
