% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rjmcmc.R
\name{mergeavg}
\alias{mergeavg}
\title{Merge blocks}
\usage{
mergeavg(sbm, edges, sbmmod, ...)
}
\arguments{
\item{sbm}{the current state of the sampler}

\item{edges}{an \code{\link{edges}} object}

\item{sbmmod}{an \code{\link{sbmmod}} model}

\item{...}{additional parameter to 'accept'}
}
\value{
an updated \code{sbm} object
}
\description{
Merge-move using an average to merge parameters
}
\details{
the blocks are chosen at random, the nodes reassigned to the block with the smallest index, then the parameters are combined using the average on the transformed scale
}
