% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rjmcmc.R
\name{mergeblocks}
\alias{mergeblocks}
\title{merge move block merging}
\usage{
mergeblocks(currblocks, propparams, edges, sbmmod, k, l)
}
\arguments{
\item{currblocks}{current blocks}

\item{propparams}{proposed parameters}

\item{edges}{an \code{\link{edges}} object}

\item{sbmmod}{an \code{\link{sbmmod}} model}

\item{k}{Blocks to merge}

\item{l}{Blocks to merge}
}
\value{
\code{list(proposed block structure, log-acceptance-prob)}
}
\description{
merge move block merging
}
