% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rjmcmc.R
\name{mergeparams.default}
\alias{mergeparams.default}
\title{Merge step: parameters}
\usage{
\method{mergeparams}{default}(params, k, l, parammod)
}
\arguments{
\item{params}{a \code{params object}}

\item{k}{Blocks to merge}

\item{l}{Blocks to merge}

\item{parammod}{a \code{parammod} object}
}
\value{
\code{list(proposed_params, log-acceptance-prob)}
}
\description{
Merge step: parameters
}
