% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rjmcmc.R
\name{mergeparams.numeric}
\alias{mergeparams.numeric}
\title{Merge step - parameter merging}
\usage{
\method{mergeparams}{numeric}(thetak, thetal, x, parammod)
}
\arguments{
\item{thetak, thetal}{parameters to merge}

\item{x}{auxiliary parameter}

\item{parammod}{a \code{parammod} object}
}
\value{
\code{list(proposed_params, log-acceptance-prob)}
}
\description{
Merge step - parameter merging
}
