% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodelike.R
\name{nodelike}
\alias{nodelike}
\title{Likelihood of node assignment}
\usage{
nodelike(blocks, params, edges, i, sbmmod, ...)
}
\arguments{
\item{blocks}{an \code{\link{blocks}} object}

\item{params}{an \code{\link{params}} object}

\item{edges}{an \code{\link{edges}} object}

\item{i}{the node of interest}

\item{sbmmod}{an \code{\link{sbmmod}} object}

\item{...}{additional arguments for \code{nodelike.blocks}}
}
\value{
likelihood of edges emanating from node i
}
\description{
Calculate the likelihood of a nod belonging to each of block
}
\details{
the number of blocks considered is either the number of blocks in \code{sbm (kappa)} or \code{kappa+1} when \code{sbmmod} has a variable number of blocks.
care is taken for data which is directed and with loops.
}
