% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traces.R
\name{numblockstrace}
\alias{numblockstrace}
\title{plot a trace of the number of blocks from MCMC samples}
\usage{
numblockstrace(postk, burnin)
}
\arguments{
\item{postk}{output from sampler}

\item{burnin}{which iterations to plot? defaults to all.}
}
\value{
`ggplot2` object
}
\description{
plot a trace of the number of blocks from MCMC samples
}
