% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parammat.R
\name{parammat.blocks}
\alias{parammat.blocks}
\title{Parameter Matrix}
\usage{
\method{parammat}{blocks}(x, params, ...)
}
\arguments{
\item{x}{a \code{blocks} object}

\item{params}{a \code{params} object}

\item{...}{(unused)}
}
\value{
an \code{NxN} matrix \code{P}, with \code{P[i,j] = } the parameter governing edge \code{ij}
}
\description{
Make a matrix of parameters from a \code{blocks} and \code{params} object
}
