% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parammat.R
\name{parammat.matrix}
\alias{parammat.matrix}
\title{Parameter Matrix}
\usage{
\method{parammat}{matrix}(zleft, zright, params, ...)
}
\arguments{
\item{zleft}{block assignment matrix on the left}

\item{zright}{block assignment matrix on the right}

\item{params}{the parameters object}

\item{...}{(unused)}
}
\value{
a matrix of parameters of size |\code{zleft}| x |\code{zright}|
}
\description{
Make a matrix of parameters from a matrix of block assignments
}
