% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parammat.R
\name{parammat.params}
\alias{parammat.params}
\title{Parameter Matrix}
\usage{
\method{parammat}{params}(x, kappa, ...)
}
\arguments{
\item{x}{a \code{params} object}

\item{kappa}{- number of blocks to compute for matrix (optional)}

\item{...}{(unused)}
}
\value{
a matrix of parameters
}
\description{
Make a matrix of parameters from a \code{params} object
}
