% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parammat.R
\name{parammat.sbm}
\alias{parammat.sbm}
\title{Parameter Matrix}
\usage{
\method{parammat}{sbm}(x, ...)
}
\arguments{
\item{x}{an \code{sbm} object}

\item{...}{(unused)}
}
\value{
a matrix of parameters
}
\description{
Make a matrix of parameters from an \code{sbm} object
}
