% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{params}
\alias{params}
\title{\code{params} S3 object}
\usage{
params(theta0, thetak)
}
\arguments{
\item{theta0}{between block parameters - a vector of length `dimension of theta`}

\item{thetak}{within block parameters - a matrix with \code{ncol=kappa} and \code{nrow=dimension of theta}}
}
\value{
a \code{params} object
}
\description{
make a \code{params} object from the between-block parameter \code{theta0} and a vector of within block parameters \code{thetak}
}
\examples{
p <- params(0.1, c(0.2,0.4,0.5))
p
}
