% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traces.R
\name{paramtrace}
\alias{paramtrace}
\title{plot a trace of parameter values from MCMC samples}
\usage{
paramtrace(theta, range, burnin)
}
\arguments{
\item{theta}{output from sampler}

\item{range}{which thetas to plot? defaults to all.}

\item{burnin}{which iterations to plot? defaults to all.}
}
\value{
`ggplot2` object
}
\description{
plot a trace of parameter values from MCMC samples
}
