% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocks.R
\name{plot.blocks}
\alias{plot.blocks}
\alias{image.blocks}
\title{Plot blocks}
\usage{
\method{plot}{blocks}(x, col, xaxt = "n", yaxt = "n", xlab = "Nodes", ylab = "Nodes", ...)

\method{image}{blocks}(x, col, xaxt = "n", yaxt = "n", xlab = "Nodes", ylab = "Nodes", ...)
}
\arguments{
\item{x}{a blocks object to plot}

\item{col}{colours for the plot}

\item{xaxt}{override \code{image} parameters}

\item{yaxt}{override \code{image} parameters}

\item{xlab}{override \code{image} parameters}

\item{ylab}{override \code{image} parameters}

\item{...}{additional parameters for \code{image}}
}
\description{
plots a block object
}
\details{
plot the block assignments in a \code{blocks} object as a matrix, color-coded by block membership
}
\examples{
## Assign six nodes to four blocks:
b <- blocks(c(1,1,2,3,4,4), 4)
plot(b)
## note that the lower left corner has one 2x2 red square
## indicating node 1 and 2 belong to the same block
}
