% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{rcat}
\alias{rcat}
\title{Draw draw Categorical distribution}
\usage{
rcat(n, p, replace = TRUE)
}
\arguments{
\item{n}{number of draws}

\item{p}{a length-d probability vector}

\item{replace}{should the categories be replaced? If so n < p required}
}
\value{
a draw from Categorical(p)
}
\description{
Draw draw Categorical distribution
}
\examples{
rcat(1, 1) ## returns 1 with probability 1
rcat(1, rep(1/6,6)) ## a dice roll
}
