% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{rdirichlet}
\alias{rdirichlet}
\title{Dirichlet distribution}
\usage{
rdirichlet(n, gam)
}
\arguments{
\item{n}{number of variates to draw}

\item{gam}{a vector of concentration parameters of length \code{K}}
}
\value{
matrix dimension \code{n*k} of samples
}
\description{
Draw from Dirichlet distribution
}
\examples{
rdirichlet(1, rep(2,5)) ## a length-5 probability vector
}
