% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawparams.R
\name{rw}
\alias{rw}
\title{Random Walk}
\usage{
rw(p, pm, sigma)
}
\arguments{
\item{p}{a parameter}

\item{pm}{a \code{parammod} object}

\item{sigma}{- scale of random walk}
}
\value{
\code{ist(proposed parameter, locjacobian)}
}
\description{
performs a random walk on a parameter value with a given parameter model
}
