% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rjmcmc.R
\name{splitavg}
\alias{splitavg}
\title{split move using average to merge parameters}
\usage{
splitavg(sbm, edges, sbmmod, ...)
}
\arguments{
\item{sbm}{the current state of the sampler}

\item{edges}{an \code{\link{edges}} object}

\item{sbmmod}{an \code{\link{sbmmod}} model}

\item{...}{additional parameter to 'accept'}
}
\value{
an updated \code{sbm} object
}
\description{
split move using average to merge parameters
}
