% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rjmcmc.R
\name{splitblocks}
\alias{splitblocks}
\title{split move: blocks}
\usage{
splitblocks(currblocks, propparams, edges, sbmmod, k)
}
\arguments{
\item{currblocks}{current blocks}

\item{propparams}{proposed parameters}

\item{edges}{an \code{edges} object}

\item{sbmmod}{a model list}

\item{k}{block to split}
}
\value{
\code{list(proposed block structure, log-acceptance-prob)}
}
\description{
split move: blocks
}
