% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rjmcmc.R
\name{splitparams.params}
\alias{splitparams.params}
\title{split move: \code{params}}
\usage{
\method{splitparams}{params}(params, k, parammod)
}
\arguments{
\item{params}{a \code{params} object to split}

\item{k}{block to split}

\item{parammod}{\code{parammod} object}
}
\value{
\code{list(proposed_params, log-acceptance-prob)}
}
\description{
split move: \code{params}
}
