% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rjmcmc.R
\name{splitparams}
\alias{splitparams}
\title{split move: parameters}
\usage{
splitparams(x, ...)
}
\arguments{
\item{x}{object for dispatch}

\item{...}{additional arguments for method}
}
\value{
\code{list(proposed_params, log-acceptance-prob)}
}
\description{
split move: parameters
}
