% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateblock.R
\name{updateblock.blocks}
\alias{updateblock.blocks}
\title{Update the block assignment of a node}
\usage{
\method{updateblock}{blocks}(blocks, i, newblock)
}
\arguments{
\item{blocks}{a \code{\link{blocks}} object}

\item{i}{the node to update}

\item{newblock}{the new block for node i}
}
\value{
new \code{blocks} object
}
\description{
change the block assignment in an \code{blocks} object to a new block
}
