% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocks.R
\name{updateblock.sbm}
\alias{updateblock.sbm}
\title{Update the block assignment of a node}
\usage{
\method{updateblock}{sbm}(sbm, newblock, i)
}
\arguments{
\item{sbm}{an sbm object}

\item{newblock}{the new block for node i}

\item{i}{the node to update}
}
\value{
new sbm object
}
\description{
Update the block assignment of a node
}
